[RENAME] module common 내 파일명, 함수명 수정하기 #NADA-49
