+++ src/components/community/comment/Profile.js
@@ -1,5 +1,5 @@
 /** Profile 프로필 컴포넌트  */
-import Dateformat from '../../../modules/common/Dateformat';
+import getDateFormat from '../../../modules/common/getDateFormat';
 import { ProfileStyle } from '../../../styles/community/CommentStyle';
 
 export const Profile = ({ comment, isEdited = comment.isEdited }) => {
@@ -9,7 +9,7 @@ export const Profile = ({ comment, isEdited = comment.isEdited }) => {
       <div className="name-date">
         <div className="name">{comment.userName}</div>
         <div className="line">
-          <div className="date">{Dateformat(comment.updatedAt)}</div>
+          <div className="date">{getDateFormat(comment.updatedAt)}</div>
           {isEdited && <div className="update">(수정됨)</div>}
         </div>
       </div>
