+++ src/containers/community/comment/BottomBar.js
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux/es/hooks/useSelector';
 import { Bar } from '../../../styles/community/CommentBarStyle';
 import usePostComment from '../../../modules/queries/community/usePostComment';
 import usePatchComment from '../../../modules/queries/community/usePatchComment';
-import preSetForQuery from '../../../modules/common/preSetForQuery';
+import preSetForQuery from '../../../modules/common/getBasicUrl';
 import NoticeBar from './NoticeBar';
 import { InputBar, Img } from '../../../styles/community/CommentBarStyle';
 import { ArrowSvg } from '../../../icon/ArrowSvg';
