+++ src/modules/common/getBasicUrl.js
@@ -1,6 +1,6 @@
-/**preSetForQuery 신고 외 모달이 열린 위치에 따라 쿼리에 사용할 api, idData를 전처리해주는 컴포넌트*/
+/**getBasicUrl 신고 외 모달이 열린 위치에 따라 쿼리에 사용할 api, idData를 전처리해주는 컴포넌트*/
 
-const preSetForQuery = (position, PostDetail, activity = '') => {
+const getBasicUrl = (position, PostDetail, activity = '') => {
   const positionsConfig = {
     activity: { url: '/activity', idData: activity._id },
     post: {
@@ -20,4 +20,4 @@ const preSetForQuery = (position, PostDetail, activity = '') => {
   return positionsConfig[position] || {};
 };
 
-export default preSetForQuery;
+export default getBasicUrl;
