+++ src/modules/common/getDateFormat.js
@@ -1,5 +1,5 @@
 /** Dateformat date를 <N년 N월 N일 N:N> 형식으로 포맷 */
-const formatDate = (preDate) => {
+const getDateFormat = (preDate) => {
   const date = new Date(preDate);
 
   const padZero = (value) => (value < 10 ? `0${value}` : value);
@@ -13,4 +13,4 @@ const formatDate = (preDate) => {
   return `${year}년 ${month}월 ${day}일 ${hours}:${minutes}`;
 };
 
-export default formatDate;
+export default getDateFormat;
