+++ src/modules/common/getReportUrl.js
@@ -1,5 +1,5 @@
-/**toReport 신고 시 모달이 열린 위치에 따라 쿼리에 사용할 api, idData를 전처리해주는 컴포넌트*/
-const toReport = (position, PostDetail) => {
+/**getReportUrl 신고 시 모달이 열린 위치에 따라 쿼리에 사용할 api, idData를 전처리해주는 컴포넌트*/
+const getReportUrl = (position, PostDetail) => {
   const positionsConfig = {
     post: {
       url: '/community/post/report',
@@ -18,4 +18,4 @@ const toReport = (position, PostDetail) => {
   return positionsConfig[position] || {};
 };
 
-export default toReport;
+export default getReportUrl;
